cssvalues: reduce CSSValues size
* switch from float64 -> float32; other browsers use 32-bit floats too
* specify integer size as 32-bit
* use NetworkBitmap for background-image value (currently just an
  invalid dummy value)
* remove "none" property & value types

CSSValue's payload is always one word (plus another for the type tag).
CSSValues keeps its size, but no longer has to heap-alloc + refcount
word-sized CSSValues.

(On 32-bit systems, CSSValues might actually be larger than before,
but I expect it's still a net benefit with the removal of refcounting
and the switch to 32-bit floats.)

+++ src/config/toml.nim
@@ -35,7 +35,7 @@ type
     of tvtInteger:
       i*: int64
     of tvtFloat:
-      f*: float64
+      f*: float32
     of tvtBoolean:
       b*: bool
     of tvtTable:
@@ -455,7 +455,7 @@ proc consumeNumber(state: var TomlParser; buf: openArray[char]; c: char):
       return state.err("invalid octal number")
     return ok(TomlValue(t: tvtInteger, i: val.get))
   of pntFloat:
-    let val = parseFloat64(repr)
+    let val = parseFloat32(repr)
     return ok(TomlValue(t: tvtFloat, f: val))
 
 proc consumeValue(state: var TomlParser; buf: openArray[char]): TomlResult
